Move error-formatting code into otter.convergence.errors - include 'reasons' in the formatted log message
