+++ otter/log/spec.py
@@ -25,7 +25,8 @@ msg_types = {
     "execute-convergence-results": (
         "Got result of {worst_status} after executing convergence"),
     "group-status-active": "Group's status is changed to ACTIVE",
-    "group-status-error": "Group's status is changed to ERROR",
+    "group-status-error":
+        "Group's status is changed to ERROR. Reasons: {reasons}",
     "launch-servers": "Launching {num_servers} servers",
     "mark-clean-success": "Marked group {scaling_group_id} clean",
     "mark-clean-failure": "Failed to mark group {scaling_group_id} clean",
